Added early_terminate_interval settings to C, Python, Matlab

+++ interfaces/python/module/codegen/files_to_generate/example.c
@@ -14,7 +14,7 @@ int main(int argc, char **argv) {
 
     // Print status
     printf("Status:                %s\n", (&workspace)->info->status);
-    printf("Number of iterations:  %d\n", (&workspace)->info->iter);
+    printf("Number of iterations:  %d\n", (int)((&workspace)->info->iter));
     printf("Objective value:       %.4e\n", (&workspace)->info->obj_val);
     printf("Primal residual:       %.4e\n", (&workspace)->info->pri_res);
     printf("Dual residual:         %.4e\n", (&workspace)->info->dua_res);
